moving socket out of pages directory since it's not a page
