+++ client/pages/backend.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react';
-import { socket } from './socket';
+import { socket } from '../utils/socket';
+import Image from 'next/image';
 
 const Backend = () => {
   const [img, setImg] = useState<string>('');
@@ -34,7 +35,7 @@ const Backend = () => {
   return (
     <>
       <div>Welcome to the backend ui!</div>
-      {img && <img id="image" alt="Received image" src={img}></img>}
+      {img && <Image id="image" alt="Received image" src={img} />}
     </>
   );
 };
